fix: separating concerns by extracting business logic from UI component into util-function normalizeWineData
